add: populate Test DB; add: unit test deleting an article with ArticleDataProvider
